+++ executor/simple.go
@@ -324,24 +324,6 @@ func (e *SimpleExec) executeFlush(s *ast.FlushStmt) error {
 }
 
 func (e *SimpleExec) executeDropStats(s *ast.DropStatsStmt) error {
-	exec := e.ctx.(sqlexec.SQLExecutor)
-	_, err := exec.Execute("begin")
-	if err != nil {
-		return errors.Trace(err)
-	}
-	id := s.Table.TableInfo.ID
-	_, err = exec.Execute(fmt.Sprintf("update mysql.stats_meta set version = %d where table_id = %d ", e.ctx.Txn().StartTS(), id))
-	if err != nil {
-		return errors.Trace(err)
-	}
-	_, err = exec.Execute(fmt.Sprintf("delete from mysql.stats_histograms where table_id = %d", id))
-	if err != nil {
-		return errors.Trace(err)
-	}
-	_, err = exec.Execute(fmt.Sprintf("delete from mysql.stats_buckets where table_id = %d", id))
-	if err != nil {
-		return errors.Trace(err)
-	}
-	_, err = exec.Execute("commit")
-	return errors.Trace(err)
+	h := sessionctx.GetDomain(e.ctx).StatsHandle()
+	return errors.Trace(h.DeleteTableStatsFromKV(s.Table.TableInfo.ID))
 }
